server: widen the BeeCounter uptime/glitch raw_json fallback to BIGINT
Migration 024 widened measurements.bee_counter_N_uptime_s and
bee_counter_N_glitch_count from INTEGER to BIGINT, because HiveTraffic wire
revision 3 made both fields 32-bit unsigned on the device and the glitch tally
is designed to arrive at exactly 4294967295 when it saturates — above the
signed 32-bit ceiling.

The read path was left behind. MEASUREMENT_SELECT_COLUMNS reads these values
back out of raw_json for rows written before the columns existed, and that
COALESCE fallback still cast the JSON text to ::integer. A value the widened
column accepts on the way in therefore raises "integer out of range" on the way
out, failing the whole SELECT rather than one field — the same failure mode the
migration was written to remove, moved from ingest to retrieval.

Cast the four fallbacks to ::bigint so they can represent the same range as the
columns they stand in for, and record why next to them.

Select-list order and arity are unchanged (137 columns), so the positional
indices the row mapper uses — 56/63/70/77 — still line up.

+++ server/measurements.py
@@ -700,14 +700,21 @@ MEASUREMENT_SELECT_COLUMNS = """
     COALESCE(bee_counter_1_ok,                NULLIF(raw_json->>'bee_counter_1_ok',                '')::boolean) AS bee_counter_1_ok,
     COALESCE(bee_counter_1_protocol_version,  NULLIF(raw_json->>'bee_counter_1_protocol_version',  '')::integer) AS bee_counter_1_protocol_version,
     COALESCE(bee_counter_1_status_flags,      NULLIF(raw_json->>'bee_counter_1_status_flags',      '')::integer) AS bee_counter_1_status_flags,
-    COALESCE(bee_counter_1_uptime_s,          NULLIF(raw_json->>'bee_counter_1_uptime_s',          '')::integer) AS bee_counter_1_uptime_s,
+    -- uptime_s and glitch_count fall back through ::bigint, not ::integer:
+    -- migration 024 widened both COLUMNS to BIGINT for HiveTraffic wire
+    -- revision 3, which made them 32-bit unsigned on the device (glitches
+    -- saturates at exactly 4294967295, above the signed 32-bit ceiling), and
+    -- this fallback has to be able to represent the same range the column it
+    -- stands in for does. An ::integer cast here raises "integer out of
+    -- range" and fails the whole SELECT, not just the one field.
+    COALESCE(bee_counter_1_uptime_s,          NULLIF(raw_json->>'bee_counter_1_uptime_s',          '')::bigint)  AS bee_counter_1_uptime_s,
     COALESCE(bee_counter_1_num_gates,         NULLIF(raw_json->>'bee_counter_1_num_gates',         '')::integer) AS bee_counter_1_num_gates,
     COALESCE(bee_counter_1_gates_healthy,     NULLIF(raw_json->>'bee_counter_1_gates_healthy',     '')::integer) AS bee_counter_1_gates_healthy,
     COALESCE(bee_counter_1_total_in,          NULLIF(raw_json->>'bee_counter_1_total_in',          '')::bigint)  AS bee_counter_1_total_in,
     COALESCE(bee_counter_1_total_out,         NULLIF(raw_json->>'bee_counter_1_total_out',         '')::bigint)  AS bee_counter_1_total_out,
     COALESCE(bee_counter_1_interval_in,       NULLIF(raw_json->>'bee_counter_1_interval_in',       '')::bigint)  AS bee_counter_1_interval_in,
     COALESCE(bee_counter_1_interval_out,      NULLIF(raw_json->>'bee_counter_1_interval_out',      '')::bigint)  AS bee_counter_1_interval_out,
-    COALESCE(bee_counter_1_glitch_count,      NULLIF(raw_json->>'bee_counter_1_glitch_count',      '')::integer) AS bee_counter_1_glitch_count,
+    COALESCE(bee_counter_1_glitch_count,      NULLIF(raw_json->>'bee_counter_1_glitch_count',      '')::bigint)  AS bee_counter_1_glitch_count,
     -- busy_retries / read_attempts / latch_succeeded are READ-ONLY history. They
     -- described the wired I2C transport — bus retries and whether CMD_LATCH
     -- landed — which no firmware speaks anymore, so ingest no longer accepts or
@@ -719,14 +726,14 @@ MEASUREMENT_SELECT_COLUMNS = """
     COALESCE(bee_counter_2_ok,                NULLIF(raw_json->>'bee_counter_2_ok',                '')::boolean) AS bee_counter_2_ok,
     COALESCE(bee_counter_2_protocol_version,  NULLIF(raw_json->>'bee_counter_2_protocol_version',  '')::integer) AS bee_counter_2_protocol_version,
     COALESCE(bee_counter_2_status_flags,      NULLIF(raw_json->>'bee_counter_2_status_flags',      '')::integer) AS bee_counter_2_status_flags,
-    COALESCE(bee_counter_2_uptime_s,          NULLIF(raw_json->>'bee_counter_2_uptime_s',          '')::integer) AS bee_counter_2_uptime_s,
+    COALESCE(bee_counter_2_uptime_s,          NULLIF(raw_json->>'bee_counter_2_uptime_s',          '')::bigint)  AS bee_counter_2_uptime_s,
     COALESCE(bee_counter_2_num_gates,         NULLIF(raw_json->>'bee_counter_2_num_gates',         '')::integer) AS bee_counter_2_num_gates,
     COALESCE(bee_counter_2_gates_healthy,     NULLIF(raw_json->>'bee_counter_2_gates_healthy',     '')::integer) AS bee_counter_2_gates_healthy,
     COALESCE(bee_counter_2_total_in,          NULLIF(raw_json->>'bee_counter_2_total_in',          '')::bigint)  AS bee_counter_2_total_in,
     COALESCE(bee_counter_2_total_out,         NULLIF(raw_json->>'bee_counter_2_total_out',         '')::bigint)  AS bee_counter_2_total_out,
     COALESCE(bee_counter_2_interval_in,       NULLIF(raw_json->>'bee_counter_2_interval_in',       '')::bigint)  AS bee_counter_2_interval_in,
     COALESCE(bee_counter_2_interval_out,      NULLIF(raw_json->>'bee_counter_2_interval_out',      '')::bigint)  AS bee_counter_2_interval_out,
-    COALESCE(bee_counter_2_glitch_count,      NULLIF(raw_json->>'bee_counter_2_glitch_count',      '')::integer) AS bee_counter_2_glitch_count,
+    COALESCE(bee_counter_2_glitch_count,      NULLIF(raw_json->>'bee_counter_2_glitch_count',      '')::bigint)  AS bee_counter_2_glitch_count,
     COALESCE(bee_counter_2_busy_retries,      NULLIF(raw_json->>'bee_counter_2_busy_retries',      '')::integer) AS bee_counter_2_busy_retries,
     COALESCE(bee_counter_2_read_attempts,     NULLIF(raw_json->>'bee_counter_2_read_attempts',     '')::integer) AS bee_counter_2_read_attempts,
     COALESCE(bee_counter_2_latch_succeeded,   NULLIF(raw_json->>'bee_counter_2_latch_succeeded',   '')::boolean) AS bee_counter_2_latch_succeeded,
